Moved tests and changed config to JSON.
