+++ scripts/test.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python
+
+if __name__ == "__main__":
+    pass
