--- test/session.regex
@@ -1,1 +1,0 @@
-(TBA|[MTWRFS]+)\s*([0-9]{4}[AP]M)-([0-9]{4}[AP]M)\s*([A-Z\-]{0,8})\s*([0-9]+[A-Z]+|[A-Z]+[0-9]+|[0-9]+)?\s*\n
