+++ tests/invalid-building.test
