+++ tests/little-spacing.test
