+++ tests/session-no-rooms.test
